requirements.txt
tensorflow
pandas-datareader
pydot
pydot-ng
yfinance
pandas
matplotlib
scikit-learn
minio
tf2onnx
kfp
kfp-tekton==1.5.4
python-dotenv